requirements.txt
absl-py==0.9.0
alabaster==0.7.12
argh==0.26.2
asn1crypto==1.3.0
astor==0.8.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
awscli==1.15.39
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.0
blinker==1.4
botocore==1.10.39
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
docutils==0.16
entrypoints==0.3
flake8==3.7.9
Flask==0.12.2
future==0.18.2
gast==0.2.2
gevent==20.6.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
greenlet==0.4.16
grpcio==1.27.2
gunicorn==19.7.1
h5py==2.10.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.11.1
jmespath==0.9.3
joblib==0.15.1
jsonify==0.5
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras==2.1.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.4
numpy==1.17.2
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.6.1
pathtools==0.1.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prompt-toolkit==3.0.3
protobuf==3.11.4
psutil==5.6.7
pyasn1==0.4.3
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==4.0.1
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyreadline==2.1
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.3
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.16.0
rsa==3.4.2
Rtree==0.9.3
s3transfer==0.1.13
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
sip==4.19.13
six==1.14.0
sklearn==0.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==4.0.1
spyder-kernels==1.8.1
statsmodels==0.11.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
ujson==1.35
urllib3==1.25.8
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.14.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
yapf==0.28.0
zipp==2.2.0
zope.event==4.4
zope.interface==4.7.1
